Update MesonToolchain and AutoToolsToolchain to use a gcc wrapper.
meson will never pass any compiler flags whilst testing the compiler
which does not work with our compilers which require a sysroot and
binutils dir to function.

Autotools will not pass cflags to the CC_FOR_BUILD compiler when testing
if it works, so our --sysroot does not make it through.