Add PkgConfigCache for reproducing data from original .pc files
PkgConfigCache allows us to read in the .pc files generated by build in def package() and store them as yaml for later reading during def package_info() where it is used to populate cpp_info in a way in which PkgConfigDeps should generate .pc files containing all the data we need.

Also improves the PkgConfig tool to clear sysroot and support setting prefix.